Refactorizamos el backend
Para que no este todo en los mismos archivos y sea más fácil de entender y leer lo he migrado a archivos diferentes pero sin tocar su funcionalidad
