+++ backend/EquusTrackBackend/Database.cs
@@ -1,10 +1,8 @@
 ﻿using MySql.Data.MySqlClient;
-using System;
-using System.Data;
 
 namespace EquusTrackBackend
 {
-    internal class Database
+    public class Database
     {
         private const string connectionString = "server=localhost;database=EquusTrackDB;user=root;password=123456789;";
 
@@ -14,248 +12,5 @@ namespace EquusTrackBackend
             connection.Open();
             return connection;
         }
-
-        public static bool Login(string email, string password)
-        {
-            using (var conn = GetConnection())
-            {
-                string query = "SELECT PasswordHash FROM Usuarios WHERE Email = @Email";
-                using var cmd = new MySqlCommand(query, conn);
-                cmd.Parameters.AddWithValue("@Email", email);
-
-                var result = cmd.ExecuteScalar();
-                if (result != null)
-                {
-                    string storedHash = result.ToString();
-                    return BCrypt.Net.BCrypt.Verify(password, storedHash);
-                }
-                return false;
-            }
-        }
-
-        public static void InsertarAdminSiNoExiste()
-        {
-            using (var conn = GetConnection())
-            {
-                string checkQuery = "SELECT COUNT(*) FROM Usuarios WHERE Email = 'admin@gestorcaballos.com'";
-                using var checkCmd = new MySqlCommand(checkQuery, conn);
-                int count = Convert.ToInt32(checkCmd.ExecuteScalar());
-
-                if (count == 0)
-                {
-                    string hash = BCrypt.Net.BCrypt.HashPassword("admin1234");
-
-                    /*string insert = @"INSERT INTO Usuarios (Nombre, Apellido, Email, PasswordHash, Rol)
-                              VALUES ('Admin', 'Principal', 'admin@gestorcaballos.com', @Hash, 'Administrador')";*/
-                    string insert = @"INSERT INTO Usuarios (Nombre, Apellido, Email, PasswordHash, Rol, FechaNacimiento, Genero) VALUES ('Admin', 'Principal', 'admin@gestorcaballos.com', @Hash, 'Administrador', '1980-01-01', 'Otro')";
-
-                    using var cmd = new MySqlCommand(insert, conn);
-                    cmd.Parameters.AddWithValue("@Hash", hash);
-                    cmd.ExecuteNonQuery();
-                    Console.WriteLine("Administrador creado con éxito.");
-                }
-                else
-                {
-                    Console.WriteLine("Administrador ya existe.");
-                }
-            }
-        }
-
-        public static bool RegistrarUsuario(string nombre, string apellido, string email, string password, string rol, DateTime fechaNacimiento, string genero)
-        {
-            if (rol != "Jinete" && rol != "Entrenador")
-            {
-                Console.WriteLine("Rol no válido.");
-                return false;
-            }
-
-            using (var conn = GetConnection())
-            {
-                // Verificar si el email ya existe
-                string checkQuery = "SELECT COUNT(*) FROM Usuarios WHERE Email = @Email";
-                using var checkCmd = new MySqlCommand(checkQuery, conn);
-                checkCmd.Parameters.AddWithValue("@Email", email);
-                int count = Convert.ToInt32(checkCmd.ExecuteScalar());
-                if (count > 0)
-                {
-                    Console.WriteLine("Ya existe un usuario con ese correo.");
-                    return false;
-                }
-
-                // Insertar nuevo usuario
-                string hash = BCrypt.Net.BCrypt.HashPassword(password);
-                string insertQuery = @"INSERT INTO Usuarios 
-                    (Nombre, Apellido, Email, PasswordHash, Rol, FechaNacimiento, Genero)
-                    VALUES (@Nombre, @Apellido, @Email, @Hash, @Rol, @FechaNacimiento, @Genero)";
-                using var insertCmd = new MySqlCommand(insertQuery, conn);
-                insertCmd.Parameters.AddWithValue("@Nombre", nombre);
-                insertCmd.Parameters.AddWithValue("@Apellido", apellido);
-                insertCmd.Parameters.AddWithValue("@Email", email);
-                insertCmd.Parameters.AddWithValue("@Hash", hash);
-                insertCmd.Parameters.AddWithValue("@Rol", rol);
-                insertCmd.Parameters.AddWithValue("@FechaNacimiento", fechaNacimiento);
-                insertCmd.Parameters.AddWithValue("@Genero", genero);
-                insertCmd.ExecuteNonQuery();
-
-                Console.WriteLine("Usuario registrado correctamente.");
-                return true;
-            }
-        }
-
-
-        public class Usuario
-        {
-            public int Id { get; set; }
-            public string Nombre { get; set; }
-            public string Apellido { get; set; }
-            public string Email { get; set; }
-            public string Rol { get; set; }
-            public DateTime FechaNacimiento { get; set; }
-            public string Genero { get; set; }
-        }
-
-        public static Usuario? ValidarLogin(string email, string password)
-        {
-            using var conn = GetConnection();
-            string query = @"SELECT Id, Nombre, Apellido, Email, PasswordHash, Rol, FechaNacimiento, Genero 
-                     FROM Usuarios WHERE Email = @Email";
-            using var cmd = new MySqlCommand(query, conn);
-            cmd.Parameters.AddWithValue("@Email", email);
-
-            using var reader = cmd.ExecuteReader();
-            if (reader.Read())
-            {
-                string hash = reader.GetString("PasswordHash");
-                if (BCrypt.Net.BCrypt.Verify(password, hash))
-                {
-                    return new Usuario
-                    {
-                        Id = reader.GetInt32("Id"),
-                        Nombre = reader.GetString("Nombre"),
-                        Apellido = reader.GetString("Apellido"),
-                        Email = reader.GetString("Email"),
-                        Rol = reader.GetString("Rol"),
-                        FechaNacimiento = reader.GetDateTime("FechaNacimiento"),
-                        Genero = reader.GetString("Genero")
-                    };
-                }
-            }
-
-            return null;
-        }
-
-        public class Caballo
-        {
-            public int Id { get; set; }
-            public string Nombre { get; set; }
-            public string Raza { get; set; }
-            public string Color { get; set; }
-            public string FotoUrl { get; set; } 
-            public int IdUsuario { get; set; }
-            public int? IdEntrenador { get; set; }
-            public DateTime? FechaNacimiento { get; set; }
-            public DateTime? FechaAdopcion { get; set; }
-        }
-
-        public static List<Caballo> ObtenerCaballosPorUsuario(int idUsuario, string rol)
-        {
-            var lista = new List<Caballo>();
-            using var conn = GetConnection();
-
-            string query = rol == "entrenador"
-                ? @"SELECT c.Id, c.Nombre, c.FotoUrl
-            FROM Caballos c
-            INNER JOIN Usuarios u ON c.IdUsuario = u.Id OR c.IdEntrenador = u.Id
-            WHERE u.Id = @Id"
-                : @"SELECT Id, Nombre, FotoUrl FROM Caballos WHERE IdUsuario = @Id";
-
-            using var cmd = new MySqlCommand(query, conn);
-            cmd.Parameters.AddWithValue("@Id", idUsuario);
-
-            using var reader = cmd.ExecuteReader();
-            while (reader.Read())
-            {
-                lista.Add(new Caballo
-                {
-                    Id = reader.GetInt32("Id"),
-                    Nombre = reader.GetString("Nombre"),
-                    FotoUrl = reader.IsDBNull("FotoUrl") ? null : reader.GetString("FotoUrl")
-                });
-            }
-
-            return lista;
-        }
-
-        public static bool CrearCaballo(int idUsuario, string nombre, string raza, string color, string fotoUrl, DateTime? fechaNacimiento, DateTime? fechaAdopcion, int? idEntrenador = null)
-        {
-            using var conn = GetConnection();
-
-            // Validar que el usuario existe
-            string checkUserQuery = "SELECT COUNT(*) FROM Usuarios WHERE Id = @IdUsuario";
-            using var checkCmd = new MySqlCommand(checkUserQuery, conn);
-            checkCmd.Parameters.AddWithValue("@IdUsuario", idUsuario);
-            int userCount = Convert.ToInt32(checkCmd.ExecuteScalar());
-
-            if (userCount == 0)
-            {
-                Console.WriteLine($"Error: El usuario con Id {idUsuario} no existe.");
-                return false;
-            }
-
-            string query = @"INSERT INTO Caballos (Nombre, Raza, Color, FotoUrl, IdUsuario, IdEntrenador, FechaNacimiento, FechaAdopcion)
-                     VALUES (@Nombre, @Raza, @Color, @FotoUrl, @IdUsuario, @IdEntrenador, @FechaNacimiento, @FechaAdopcion)";
-
-            using var cmd = new MySqlCommand(query, conn);
-            cmd.Parameters.AddWithValue("@Nombre", nombre);
-            cmd.Parameters.AddWithValue("@Raza", raza ?? (object)DBNull.Value);
-            cmd.Parameters.AddWithValue("@Color", color ?? (object)DBNull.Value);
-            cmd.Parameters.AddWithValue("@FotoUrl", fotoUrl ?? (object)DBNull.Value);
-            cmd.Parameters.AddWithValue("@IdUsuario", idUsuario);
-            cmd.Parameters.AddWithValue("@IdEntrenador", idEntrenador ?? (object)DBNull.Value);
-
-            // Para fechas, usar DBNull si son null
-            cmd.Parameters.AddWithValue("@FechaNacimiento", fechaNacimiento ?? (object)DBNull.Value);
-            cmd.Parameters.AddWithValue("@FechaAdopcion", fechaAdopcion ?? (object)DBNull.Value);
-
-            int filasAfectadas = cmd.ExecuteNonQuery();
-            return filasAfectadas > 0;
-        }
-
-        public static Caballo? ObtenerCaballoPorId(int id)
-        {
-            using var conn = GetConnection(); // Asegúrate de usar el método correcto
-            string query = @"SELECT Id, Nombre, Raza, Color, FotoUrl, IdUsuario, IdEntrenador, FechaNacimiento, FechaAdopcion 
-                     FROM Caballos WHERE Id = @Id";
-
-            using var cmd = new MySqlCommand(query, conn);
-            cmd.Parameters.AddWithValue("@Id", id);
-
-            using var reader = cmd.ExecuteReader();
-            if (reader.Read())
-            {
-                return new Caballo
-                {
-                    Id = reader.GetInt32("Id"),
-                    Nombre = reader.GetString("Nombre"),
-                    Raza = reader.GetString("Raza"),
-                    Color = reader.GetString("Color"),
-                    FotoUrl = reader.GetString("FotoUrl"),
-                    IdUsuario = reader.GetInt32("IdUsuario"),
-                    IdEntrenador = reader.IsDBNull(reader.GetOrdinal("IdEntrenador"))
-                ? null
-                : reader.GetInt32("IdEntrenador"),
-                    FechaNacimiento = reader.IsDBNull(reader.GetOrdinal("FechaNacimiento"))
-                ? null
-                : reader.GetDateTime("FechaNacimiento"),
-                    FechaAdopcion = reader.IsDBNull(reader.GetOrdinal("FechaAdopcion"))
-                ? null
-                : reader.GetDateTime("FechaAdopcion")
-                };
-            }
-
-            return null;
-        }
-
-
     }
 }
