+++ backend/EquusTrackBackend/Models/Requests/CaballosRequest.cs
@@ -1,0 +1,9 @@
+﻿
+namespace EquusTrackBackend.Models.Requests
+{
+    public class CaballosRequest
+    {
+        public int IdUsuario { get; set; }
+        public required string Rol { get; set; }
+    }
+}
