+++ backend/EquusTrackBackend/Models/Requests/LoginRequest.cs
@@ -1,0 +1,9 @@
+﻿
+namespace EquusTrackBackend.Models.Requests
+{
+    public class LoginRequest
+    {
+        public required string Email { get; set; }
+        public required string Password { get; set; }
+    }
+}
