+++ backend/EquusTrackBackend/Utils/Helpers.cs
@@ -1,0 +1,40 @@
+﻿using System.Net;
+using System.Text.Json;
+
+namespace EquusTrackBackend.Utils
+{
+    public static class Helpers
+    {
+        public static void AgregarCabecerasCORS(HttpListenerResponse response)
+        {
+            response.AddHeader("Access-Control-Allow-Origin", "http://localhost:5173");
+            response.AddHeader("Access-Control-Allow-Methods", "POST, GET, OPTIONS");
+            response.AddHeader("Access-Control-Allow-Headers", "Content-Type");
+        }
+
+        public static void EnviarOpcionesCORS(HttpListenerResponse response)
+        {
+            response.StatusCode = 200;
+            AgregarCabecerasCORS(response);
+            response.Close();
+        }
+
+        public static async Task EnviarErrorRespuesta(HttpListenerContext context, Exception ex, string mensaje)
+        {
+            Console.WriteLine($"[ERROR] {mensaje}: {ex.Message}");
+            context.Response.StatusCode = 500;
+            context.Response.ContentType = "application/json";
+            AgregarCabecerasCORS(context.Response);
+
+            using var writer = new StreamWriter(context.Response.OutputStream);
+            await writer.WriteAsync(JsonSerializer.Serialize(new
+            {
+                exito = false,
+                mensaje,
+                detalle = ex.Message
+            }));
+            await writer.FlushAsync();
+            context.Response.Close();
+        }
+    }
+}
